Task2/Plant Diseases/requirements.txt
streamlit
tensorflow
pillow
numpy
requests